CleanProject/requirements.txt
alembic==1.10.2
aniso8601==9.0.1
black==23.1.0
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.6
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
